Make sure TypeExtensions/ReflectionExtensions code keeps working.
This commit brings up the last of the legacy
Reflection api that underlie the S.R.ReflectionExtensions
and S.R.TypeExtensions extension methods. This will prevent
C# code that used to compile against those libraries
from breaking once they start compiling against the new
System.Runtime contract (thanks to C#'s disambiguation
rules regarding extension methods vs. instance methods,
that recompiled code will no longer go through the extension
libraries after that recompile.)

Once the new System.Runtime libraries are ready,
we'll be free to get rid of most(*) of the AOT-specific
code for ReflectionExtension and TypeExtensions
runtime libraries and use the same code we do for
CoreClr.

(*) HasMetadataToken() and HasModuleVersionId()
have no underlying apis to call - rather, they hard-code
"true" or "false" for their respective platforms.

All Reflection tests in ToF now pass after overwriting
the AOT-specific versions of these libraries with home-cooked
versions of the above libraries (linking directly to S.P.Corelib
since the new S.Runtime library isn't ready yet.)

Notes:

  The RuntimeReflectionExtensions class itself is not
  in S.P.Corelib. Unless there's a good argument otherwise,
  it can go into the the S.Runtime partial facade (along
  with the other stuff that's there like the Action
  delegages with 16 generic parameters.)

  CustomAttributeExtensions *is* in S.P.Corelib
  since that's actually our main implementation
  for instance-based custom attribute support
  in .Net Native. All the others thunk to it rather
  than vice-versa.
