+++ src/System.Private.Reflection.Core/src/System/Reflection/Runtime/ParameterInfos/RuntimeParameterInfo.cs
@@ -43,6 +43,14 @@ namespace System.Reflection.Runtime.ParameterInfos
 
         public abstract override bool HasDefaultValue { get; }
 
+        public sealed override int MetadataToken
+        {
+            get
+            {
+                throw new InvalidOperationException(SR.NoMetadataTokenAvailable);
+            }
+        }
+
         public sealed override MemberInfo Member
         {
             get
