+++ src/System.Private.Reflection.Core/src/System/Reflection/Runtime/TypeInfos/RuntimeTypeInfo.cs
@@ -371,13 +371,33 @@ namespace System.Reflection.Runtime.TypeInfos
             return cachedData.GetDeclaredProperty(name);
         }
 
-        public sealed override Type[] GetGenericArguments()
+        public sealed override MemberInfo[] GetDefaultMembers()
         {
-            if (IsConstructedGenericType)
-                return GenericTypeArguments;
-            if (IsGenericTypeDefinition)
-                return GenericTypeParameters;
-            return Array.Empty<Type>();
+            Type defaultMemberAttributeType = typeof(DefaultMemberAttribute);
+            for (Type type = this; type != null; type = type.BaseType)
+            {
+                foreach (CustomAttributeData attribute in type.CustomAttributes)
+                {
+                    if (attribute.AttributeType == defaultMemberAttributeType)
+                    {
+                        // NOTE: Neither indexing nor cast can fail here. Any attempt to use fewer than 1 argument
+                        // or a non-string argument would correctly trigger MissingMethodException before
+                        // we reach here as that would be an attempt to reference a non-existent DefaultMemberAttribute
+                        // constructor.
+                        Debug.Assert(attribute.ConstructorArguments.Count == 1 && attribute.ConstructorArguments[0].Value is string);
+
+                        string memberName = (string)(attribute.ConstructorArguments[0].Value);
+                        return GetMember(memberName);
+                    }
+                }
+            }
+
+            return Array.Empty<MemberInfo>();
+        }
+
+        public sealed override InterfaceMapping GetInterfaceMap(Type interfaceType)
+        {
+            throw new PlatformNotSupportedException(SR.PlatformNotSupported_InterfaceMap);
         }
 
         //
@@ -504,14 +524,6 @@ namespace System.Reflection.Runtime.TypeInfos
             }
         }
 
-        public sealed override bool IsGenericType
-        {
-            get
-            {
-                return IsConstructedGenericType || IsGenericTypeDefinition;
-            }
-        }
-
         //
         // Left unsealed as generic type definitions must override.
         //
@@ -523,11 +535,11 @@ namespace System.Reflection.Runtime.TypeInfos
             }
         }
 
-        public sealed override bool IsSerializable
+        public sealed override int MetadataToken
         {
             get
             {
-                return 0 != (this.Attributes & TypeAttributes.Serializable);
+                throw new InvalidOperationException(SR.NoMetadataTokenAvailable);
             }
         }
