+++ src/System.Private.Reflection.Execution/src/System.Private.Reflection.Execution.csproj
@@ -104,9 +104,6 @@
 
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="..\..\Common\src\Internal\Reflection\Execution\Binder\DefaultBinder.cs" >
-      <Link>Internal\Reflection\Execution\Binder\DefaultBinder.cs</Link>
-    </Compile>
     <Compile Include="..\..\Common\src\Internal\Runtime\MetadataBlob.cs" >
       <Link>Internal\Runtime\MetadataBlob.cs</Link>
     </Compile>
